TST: move mutually_broadcastable_shapes from sparse to _lib
